backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "clerk-backend-api>=5.0.6",
    "fastapi>=0.135.3",
    "google-genai>=1.72.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "uvicorn>=0.43.0",
]
